backend/app/requirements.txt
flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2
python-dotenv
